+++ src/MailtrapGeneralClient.php
@@ -11,6 +11,7 @@ namespace Mailtrap;
  * @method Api\General\Contact       contacts(int $accountId)
  * @method Api\General\EmailTemplate emailTemplates(int $accountId)
  * @method Api\General\Billing       billing(int $accountId)
+ * @method Api\General\ApiToken      apiTokens(int $accountId)
  *
  * Class MailtrapGeneralClient
  */
@@ -23,5 +24,6 @@ final class MailtrapGeneralClient extends AbstractMailtrapClient
         'contacts' => Api\General\Contact::class,
         'emailTemplates' => Api\General\EmailTemplate::class,
         'billing' => Api\General\Billing::class,
+        'apiTokens' => Api\General\ApiToken::class,
     ];
 }